deployer file jenkins bbbbbbb

+++ Jenkinsfile
@@ -115,25 +115,48 @@ peCJp1UDhKUAAAAUamVua2luc0B1YnVudHUtZm9jYWwBAgMEBQYH
         }
     }
 }
+stage('Vérifier les commandes') {
+    steps {
+        script {
+            sh '''
+                # Vérifier si la commande "fin" existe
+                if ! command -v fin &> /dev/null; then
+                    echo "Erreur : la commande 'fin' n'est pas installée."
+                    exit 1
+                fi
+            '''
+        }
+    }
+}
 stage('Déploiement') {
     steps {
         script {
             sh '''
                 set -euxo pipefail
                 echo "Déploiement de l'application"
-                ssh -o StrictHostKeyChecking=no -i ~/.ssh/id_rsa vagrant@192.168.182.200 << EOF
+                ssh -o StrictHostKeyChecking=no -i ~/.ssh/id_rsa vagrant@192.168.182.200 <<EOF
                 set -euxo pipefail
-                echo "🛠️ Arrêt et suppression de l'ancien conteneur"
-                docker stop aston_villa || true
-                docker rm aston_villa || true
+                # Arrêt et suppression de l'ancien conteneur
+                if docker ps -a | grep -q "aston_villa"; then
+                    echo "🛠️ Arrêt du conteneur existant"
+                    docker stop aston_villa || true
+                    docker rm aston_villa || true
+                else
+                    echo "ℹ️ Aucun conteneur existant à supprimer"
+                fi
+
+                # Lancement du nouveau conteneur
                 echo "🚀 Lancement du nouveau conteneur"
                 docker run -d --name aston_villa -p 50:50 nawreswear/aston_villa:latest
+
                 echo "✅ Déploiement terminé avec succès."
                 EOF
-            '''
-        }
-    }
-}
+                            '''
+                        }
+
+
+                    }
+                }
